requirements-dev.txt
django-debug-toolbar

django-dynamic-fixture
fake-factory

mock
pytest
pytest-cov
pytest-django
